refactor(lsp): separate test-only helpers
Remove test-only wrappers from production modules. Route tests
through configurable production constructors and move integration
coverage into dedicated test modules.
