+++ crates/lsp/src/flycheck/config.rs
@@ -145,13 +145,7 @@ fn forge_lint_available(command: &Path, cwd: &Path) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{LaunchConfig, global_state::GlobalState, test_support::TestProject};
-    #[cfg(unix)]
-    use solar_interface::source_map::{FileLoader, RealFileLoader};
-    #[cfg(unix)]
-    use std::{fs, os::unix::fs::PermissionsExt, sync::Arc, time::Duration};
-    #[cfg(unix)]
-    use tokio::sync::oneshot;
+    use crate::test_support::TestProject;
 
     #[test]
     fn configured_flychecks_expand_per_workspace() {
@@ -199,65 +193,5 @@ mod tests {
             FlycheckInitializationOptions { forge_path: None, flychecks: Some(Vec::new()) };
 
         assert!(options.configs(project.config().workspaces()).is_empty());
-    }
-
-    #[cfg(unix)]
-    #[tokio::test(flavor = "current_thread")]
-    async fn launch_default_drives_forge_lint_discovery_and_execution() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-            [profile.default]
-            src = "src"
-
-            //- /src/Test.sol
-            contract Test {}
-            "#,
-        );
-        project.write_file(
-            "/embedded-forge",
-            r#"#!/bin/sh
-set -eu
-printf '%s\n' "$@" >> "$0.args"
-printf '%s\n' -- >> "$0.args"
-printf '%s\n' "$PWD" >> "$0.cwd"
-"#,
-        );
-        let forge = project.path("/embedded-forge");
-        let mut permissions = fs::metadata(&forge).unwrap().permissions();
-        permissions.set_mode(0o755);
-        fs::set_permissions(&forge, permissions).unwrap();
-
-        let mut state = GlobalState::with_launch_config(
-            async_lsp::ClientSocket::new_closed(),
-            LaunchConfig::default().with_default_forge_path(&forge),
-        );
-        state.on_initialize(project.initialize_params()).await.unwrap();
-        let _ = Arc::make_mut(&mut state.config).rediscover_workspaces();
-        let [config] =
-            state.config.flychecks_for_path(&project.path("/src/Test.sol")).try_into().unwrap();
-        let expected_cwd = RealFileLoader.canonicalize_path(&config.cwd).unwrap();
-
-        assert_eq!(config.command, forge);
-        assert_eq!(config.args, ["lint", "--json"]);
-        let (_cancel, cancelled) = oneshot::channel();
-        let diagnostics = crate::flycheck::run(
-            config,
-            Duration::from_secs(5),
-            cancelled,
-            vec![project.path("/src/Test.sol")],
-        )
-        .await
-        .unwrap();
-
-        assert!(diagnostics.is_empty());
-        assert_eq!(
-            project.read_file("/embedded-forge.args"),
-            "lint\n--help\n--\nlint\n--json\n--\n"
-        );
-        assert_eq!(
-            project.read_file("/embedded-forge.cwd"),
-            format!("{0}\n{0}\n", expected_cwd.display())
-        );
     }
 }
