+++ crates/lsp/src/tests/protocol_trace.rs
@@ -1,7 +1,7 @@
 use crate::{
     LaunchConfig,
     global_state::GlobalState,
-    new_router_with_state, new_server_service, new_server_service_with_router,
+    new_router_with_state, new_server_service_with_launch_config, new_server_service_with_router,
     test_support::{assert_request_cancelled, read_lsp_frame, start_request, write_lsp_frame},
 };
 use async_lsp::{
@@ -218,7 +218,9 @@ where
 }
 
 fn protocol_trace_harness() -> ProtocolTraceHarness {
-    protocol_trace_harness_with(new_server_service)
+    protocol_trace_harness_with(|client| {
+        new_server_service_with_launch_config(client, LaunchConfig::default())
+    })
 }
 
 fn protocol_trace_test_harness(pending: Option<PendingTraceControl>) -> ProtocolTraceHarness {
